--- tests/test_users.py
@@ -1,23 +1,0 @@
-import sys
-import os
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-
-from fastapi.testclient import TestClient
-from app.main import app
-
-client = TestClient(app)
-
-def test_signup_status_code():
-    response = client.post("/auth/signup", json={
-        "name": "Test User",
-        "email": "testuser@example.com",
-        "password": "secure123"
-    })
-    assert response.status_code in [200, 201, 400, 409] 
-
-def test_signin_status_code():
-    response = client.post("/auth/signin", json={
-        "email": "testuser@example.com",
-        "password": "secure123"
-    })
-    assert response.status_code in [200, 401, 404]
